requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
opencv-python==4.8.1.78
Pillow==10.1.0
PyYAML==6.0.1
tqdm==4.66.1
kagglehub==0.3.12
jupyter==1.0.0
ipykernel==6.27.1
gradio>=4.0.0
mediapipe>=0.10.0
